test: move get quote to delayed
