+++ scripts/delayed.js
@@ -1,8 +1,11 @@
 // eslint-disable-next-line import/no-cycle
 import { loadScript, sampleRUM } from './lib-franklin.js';
+import { getCookie, setCookie } from './scripts.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
+
+let refresh = false;
 
 /* eslint-disable */
 // add more delayed functionality here
@@ -47,6 +50,55 @@ function loadAccessibe() {
 }
 // Accessibe - end
 
+// Get quote - start
+async function getQuote() {
+  const reqHeaders = new Headers();
+  if (localStorage.getItem('authToken')) {
+    reqHeaders.append('Authorization', `Bearer ${localStorage.getItem('authToken')}`);
+  } else if (getCookie('ProfileData')) {
+    const { customer_token: apiToken } = getCookie('ProfileData');
+    reqHeaders.append('authentication-token', apiToken);
+  } else if (getCookie('apiToken')) {
+    const apiToken = getCookie('apiToken');
+    reqHeaders.append('authentication-token', apiToken);
+  } else if (!refresh) {
+    refresh = true;
+    const formData = 'grant_type=anonymous&scope=openid+profile&client_id=';
+    const authRequest = await fetch(`${baseURL}/token`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+      body: formData,
+    });
+    if (authRequest.ok) {
+      const data = await authRequest.json();
+      const expiresIn = data.expires_in * 1000;
+      setCookie('apiToken', data.access_token, expiresIn, '/');
+      reqHeaders.append('authentication-token', data.access_token);
+      localStorage.setItem('refreshToken', data.refresh_token);
+    }
+  }
+
+  if (reqHeaders.has('authentication-token') || reqHeaders.has('Authorization')) {
+    const quoteRequest = await fetch(`${baseURL}/rfqcart/-`, { headers: reqHeaders });
+    if (quoteRequest.ok) {
+      const data = await quoteRequest.json();
+      if (data && data.items) {
+        const rfqQuantity = data.items.length;
+        if (rfqQuantity !== 0) {
+          const quantityElement = document.querySelector('header a.quote span.quantity');
+          if (quantityElement) quantityElement.textContent = rfqQuantity;
+          const dotElement = document.querySelector('header a.quote span.dot');
+          if (dotElement) dotElement.classList.remove('hidden');
+        }
+      }
+    } else if (quoteRequest.status !== 404) {
+      // eslint-disable-next-line no-console
+      console.warn('Failed to load quote cart');
+    }
+  }
+}
+// Get quote - end
+
 // coveo analytics - start
 (function (c, o, v, e, O, u, a) {
   a = 'coveoua';
@@ -76,6 +128,7 @@ const organizationId = window.DanaherConfig !== undefined
   : 'danahernonproduction1892f3fhz';
 // coveo analytics - end
 
+getQuote();
 if (
   !window.location.hostname.includes('localhost')
   && !document.location.hostname.includes('.hlx.page')
